Minor refactor to bad request error handling, fix typo in extensions package, fix repo calls in tests to include ignoreCase.
